Updated for 0.22.  Version 1.0 release!

+++ KerbalEconomy/Ledger/Book.cs
@@ -45,7 +45,7 @@ namespace KerbalEconomy.Ledger
         /// <summary>
         /// Adds a new ledger row.
         /// </summary>
-        public void AddRow(double universalTime, string transaction, double debit, double credit, double balance)
+        public void AddRow(double universalTime, string transaction, float debit, float credit, float balance)
         {
             rows.Add(new Row()
             {
